requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
pylast==3.3.0
pytube3==9.6.4
requests==2.24.0
six==1.15.0
spotipy==2.15.0
typing-extensions==3.7.4.3
urllib3==1.25.10
